requirements.txt
requests
pymongo
pymongo
urllib.parse
matplotlib.pyplot
numpy
logging
